Check for the existence of original mlan.ko

+++ recipes-connectivity/murata-binaries/murata-binaries/switch_module_imx8mnea-ucom.sh
@@ -129,13 +129,15 @@ function prepare_for_nxp_sdio() {
   ln -s /usr/sbin/hostapd.nxp /usr/sbin/hostapd
   ln -s /usr/sbin/hostapd_cli.nxp /usr/sbin/hostapd_cli
   
-  #Remove existing link of mlan.ko
+  # Remove existing link of mlan.ko
   if [ -L /lib/modules/$(uname -r)/extra/mlan.ko ]; then
     rm /lib/modules/$(uname -r)/extra/mlan.ko
   fi
   
-  #Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  # Bring back mlan.ko from backup to extra
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
 
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
@@ -174,8 +176,9 @@ function prepare_for_nxp_xk_sdio() {
   fi
   
   # Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
-
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
 
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
@@ -214,7 +217,9 @@ function prepare_for_nxp_ds_sdio() {
   fi
   
   # Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
 
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
@@ -253,7 +258,9 @@ function prepare_for_nxp_ym_sdio() {
   fi
   
   # Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
 
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
@@ -298,7 +305,9 @@ function prepare_for_nxp_ym_pcie() {
   fi
   
   # Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
 
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
@@ -342,7 +351,9 @@ function prepare_for_nxp_xl_sdio() {
   fi
   
   # Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
 
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
@@ -386,7 +397,9 @@ function prepare_for_nxp_xl_pcie() {
   fi
   
   # Bring back mlan.ko from backup to extra
-  cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  if [ -e /usr/share/murata_wireless/mlan.ko.orig ]; then
+    cp /usr/share/murata_wireless/mlan.ko.orig /lib/modules/$(uname -r)/extra/mlan.ko
+  fi
   
   cat <<EOT > /etc/depmod.d/nxp_depmod.conf
 # Force modprobe to search kernel/net/wireless (where the NXP
